pyproject.toml
[project]
name = "yolo-fine-tuning-pipeline"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=11.0.0",
    "pyyaml>=6.0.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "trackio>=0.22.0",
    "ultralytics>=8.4.37",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
